Rename and unify files
